+++ tools/mipsdec/symbols.h
@@ -18,7 +18,7 @@ public:
 	static bool parseSymFile(const std::string &strSymFile);
 	static bool lookup(const std::string &strSymName, unsigned &uSymIdx);
 	static bool lookup(unsigned uAddress, unsigned &uSymIdx);
-	static unsigned getCount(void);
+	static size_t getCount(void);
 	static const tSymbolEntry *get(unsigned uSymIdx);
 
 private:
